Moving 2D vector operations into Vector2D shape class
