+++ src/main/java/com/spatial4j/core/shape/Vector2D.java
@@ -25,8 +25,8 @@ public class Vector2D {
     /**
      * Store 2 doubles as vector components
      */
-    private final double X;
-    private final double Y;
+    private double X;
+    private double Y;
 
     /**
      * Constructors
@@ -88,5 +88,111 @@ public class Vector2D {
         if ( thiz.getY() != v.getY() ) return false;
 
         return true;
+    }
+
+    /**
+     * Compute the sum of two vectors
+     */
+    public static Vector2D add( Vector2D a, Vector2D b ) {
+
+        double sum_x = a.getX() + b.getX();
+        double sum_y = a.getY() + b.getY();
+
+        return new Vector2D( sum_x, sum_y );
+    }
+
+    /**
+     * Compute the difference between two vectors
+     */
+    public static Vector2D subtract( Vector2D a, Vector2D b ) {
+
+        double diff_x = a.getX() - b.getX();
+        double diff_y = a.getY() - b.getY();
+
+        return new Vector2D( diff_x, diff_y );
+    }
+
+
+    /**
+     * Compute the product of a vector and a scalar factor s.
+     */
+    public static Vector2D multiply( Vector2D v, double s ) {
+
+        double mult_x = s * v.getX();
+        double mult_y = s * v.getY();
+
+        return new Vector2D( mult_x, mult_y );
+    }
+
+    /**
+     * Compute the scalar magnitude of a vector
+     */
+    public double length() {
+
+        double x_2 = Math.pow( this.X, 2 );
+        double y_2 = Math.pow( this.Y, 2 );
+
+        return Math.sqrt( x_2 + y_2 );
+    }
+
+    /**
+     * Compute the unit vector of the given vector
+     */
+    public Vector2D unitVector() {
+        return multiply(this, 1/length());
+    }
+
+    /**
+     * Compute the cross product between two 2D vectors (apparently this is mathematically valid??)
+     * returns a scalar in 2D
+     */
+    public double crossProduct( Vector2D v ) {
+        return getX()*v.getY() - getY()*v.getX();
+    }
+
+    /**
+     * Compute the dot product between two vectors
+     */
+    public double dotProduct( Vector2D v ) {
+
+        double x = getX()*v.getX();
+        double y = getY()*v.getY();
+
+        return x + y;
+    }
+
+
+    /**
+     * Standard Normalize
+     */
+    public static Vector2D normalize( Vector2D a ) {
+        double n = norm(a);
+        if ( n != 0 ) n = 1.0/n;
+        return multiply(a, n);
+    }
+
+    /**
+     * Norm2
+     */
+    public static double norm2( Vector2D a ) {
+        return Math.pow(a.getX(), 2) + Math.pow(a.getY(), 2);
+    }
+
+    /**
+     * Norm
+     */
+    public static double norm( Vector2D a ) {
+        return Math.sqrt(norm2(a));
+    }
+
+    /**
+     * Vector Comparison: B greater than A
+     */
+    public static boolean greaterThan( Vector2D a, Vector2D b ) {
+        if ( a.getX() < b.getX() ) return true;
+        if ( b.getX() < a.getX() ) return false;
+        if ( a.getY() < b.getY() ) return true;
+        return false;
+
     }
 }
